Separate out slot_worker.rs
Also make SubspaceSlotWorker `pub(super)`.
